// src/Components/StyledComponents/Types.ts
export type T_Flex = {
  flexDirection: string;
  justifyContent: string;
  alignItems: string;
  display: string;
};

export type T_Width = {
  width: string;
  maxWidth: string;
};

export type T_Height = {
  height: string;
};

export type T_Gap = {
  gap: string;
};

export type T_BoxSizing = {
  boxSizing: string;
};

export type T_Padding = {
  padding: string;
};

export type T_Margin = {
  marginT?: string;
};